dkfriend-backend-latest/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.2
django-admin==2.0.2
django-admin-thumbnails==0.2.8
django-allauth==0.54.0
django-cors-headers==4.2.0
django-excel-response2==3.0.5
django-phonenumber-field==7.1.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
excel-base==1.0.4
idna==3.4
isoweek==1.3.3
oauthlib==3.2.2
phonenumbers==8.13.16
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
xlwt==1.3.0
